backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
